MCP/Ecommerce_MCP_Server/mcp_server/requirements.txt
# MCP SDK (official MCP package for streamable-http transport)
mcp>=1.0.0

# FastMCP framework (required by AgentCore Runtime)
fastmcp>=0.1.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Authentication and JWT handling
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
cryptography>=42.0.0

# HTTP requests (for Cognito token validation)
requests>=2.31.0

# Data validation
pydantic>=2.5.0

# Environment variables
python-dotenv>=1.0.0

# Date utilities
python-dateutil>=2.8.2

# Web framework (FastMCP is built on top of FastAPI/Starlette)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
